# perf: 複数のボトルネック箇所に対処