やや小粒だがボトルネックとなっていた箇所を特定し、集中的に最適化した。

* `setUnreachable`の判定処理が重かったので、軽量化（ZenzaiTests.testGradualConversion_Roman2Kana: 5.99s -> 5.83s）
* `getWiseDicdata`の日本語数字パースが重かったので、軽量化（ZenzaiTests.testGradualConversion_Roman2Kana: 5.83s -> 5.35s）
* `getDicdataFromLoudstxt3`のパース処理が重かったので、軽量化（ZenzaiTests.testGradualConversion_Roman2Kana: 5.35s -> 4.77s）

これにより #249 で取り扱ったKanaKanjiConverterModuleWithDefaultDictionaryTests.ZenzaiTests/testGradualConversion_Roman2Kanaのケースでは、全体で 5.99s -> 4.77s（20%改善）となった。一連の最適化で、トータルでは55%程度の改善となった。

* #249 適用前: 10.46s
* #249 適用後: 7.69s
* #250 適用後: 6.71s
* #252 適用後: 5.99s
* #253 適用後: 4.77s